start splitting into pieces
